cmd: Move policy and enabled checks to `policy` package
